Pipeline redesigned to utilize the SCons build system wrapper extention of Nestly.
After discussions with the developers of Nestly and a review of other workflow tools (namely Swift) it was decided to stay with Nestly. The SCons
wrapper enables defining sources and targets (dependencies) as well as moving within the hierarchy of the parameter sweep (nest). Moving within the
tree has not been fully exploited as yet.

A few pieces of short-running python code are not being executed on the queue. (fix me)

Currently only recall, precision and f1 are being calculated.

Matt.